refactor(agent): extract workspace-encryption/sender-key into mixin holder
Move recipient/gate resolution, on-chain access-policy reads, SWM
sender-key epoch creation+distribution, pending-package queueing,
workspace payload encrypt/decrypt, and sender-key state persistence
(~2k lines) into dkg-agent-crypto.ts.
